cloudfunctions/src/cf_extract_anac_oferta_demanda_aerea_basica_to_bronze/source_code/requirements.txt
google-cloud-storage==2.15.0
google-cloud-bigquery==3.18.0
beautifulsoup4==4.12.3
requests==2.31.0
unidecode==1.3.8
loguru==0.7.2
pandas==2.2.1
pyarrow==15.0.0
pytz==2024.1
openpyxl==3.1.2
pyyaml==6.0.1